Moved NetworkWalker to another package
